+++ site/components/pages/Pledge/lib/index.tsx
@@ -1,4 +1,3 @@
 export { getPledgeByAddress, findOrCreatePledge } from "./firebase";
-export { formSchema } from "./formSchema";
 export { pledgeResolver } from "./pledgeResolver";
 export { verifySignature } from "./verifySignature";
