+++ site/components/pages/Pledge/lib/verifySignature.ts
@@ -1,4 +1,4 @@
-import { ethers } from 'ethers'
+import { ethers } from "ethers";
 
 interface Params {
   address: string;
@@ -7,9 +7,12 @@ interface Params {
 }
 
 export const verifySignature = (params: Params): void => {
-  const decodedAddress = ethers.utils.verifyMessage(params.nonce, params.signature)
+  const decodedAddress = ethers.utils.verifyMessage(
+    params.nonce,
+    params.signature
+  );
 
-  if (decodedAddress.toLowerCase() !== params.address.toLowerCase()) { 
-    throw new Error('Invalid signature')
+  if (decodedAddress.toLowerCase() !== params.address.toLowerCase()) {
+    throw new Error("Invalid signature");
   }
-}
+};
